Moved Thesaurus-related classes into their own namespace.
