+++ Logophi.csproj
@@ -124,14 +124,14 @@
     </Compile>
     <Compile Include="GlobalHotkey.cs" />
     <Compile Include="HotkeyEventArgs.cs" />
-    <Compile Include="IBookmarkManager.cs" />
+    <Compile Include="Thesaurus\IBookmarkManager.cs" />
     <Compile Include="MessageReceiver.cs" />
     <Compile Include="NativeMethods.cs" />
     <Compile Include="PresentationAgent.cs" />
-    <Compile Include="SearchHistory.cs" />
-    <Compile Include="TermEntry.cs" />
-    <Compile Include="Thesaurus.cs" />
-    <Compile Include="TunaInterface.cs" />
+    <Compile Include="Thesaurus\SearchHistory.cs" />
+    <Compile Include="Thesaurus\TermEntry.cs" />
+    <Compile Include="Thesaurus\Browser.cs" />
+    <Compile Include="Thesaurus\TunaInterface.cs" />
     <Compile Include="Views\IPreferencesDialogView.cs" />
     <Compile Include="Views\IView.cs" />
     <Compile Include="Views\Presentation\BookmarksFormPresenter.cs" />
@@ -153,7 +153,7 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Views\Presentation\PreferencesDialogPresenter.cs" />
-    <Compile Include="WordDefinition.cs" />
+    <Compile Include="Thesaurus\WordDefinition.cs" />
     <EmbeddedResource Include="Forms\AboutForm.resx">
       <DependentUpon>AboutForm.cs</DependentUpon>
     </EmbeddedResource>
