+++ Thesaurus/WordDefinition.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json;
 
-namespace MouseNet.Logophi
+namespace MouseNet.Logophi.Thesaurus
 {
     [JsonObject(MemberSerialization.OptIn), Serializable]
     public class WordDefinition
