+++ Views/Presentation/BookmarksFormPresenter.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using MouseNet.Logophi.Thesaurus;
 
 namespace MouseNet.Logophi.Views.Presentation
 {
